wip: the resize — int is i32, float is binary32
Step 4 of docs/TYPES.md.  **The mechanism is complete and every
bundled program and benchmark compiles; the executable specification
has not been migrated, so this commit is not green.**  It is recorded
on `impl-resize` rather than main for exactly that reason;
docs/TYPES.md's *As built* names the four things still owed and the
order to do them in.

**The order that made it safe.**  `types.Type.int` became `.long` and
`.float` became `.double` tree-wide *first* — with `Type.Payload`,
`runtime.Value.Tag` (2 and 3 keeping their numbers), `runtime.View`,
`heap.ElementKind`, `context.ConstantValue`, `lower.Numeric` and the
`ofInt`/`asInt`/`ofFloat`/`asFloat` accessors — and only then did
`int` and `float` come back at 32 bits.  There is no commit in which
either name means the wrong width, and no site can have been missed,
because a missed one does not compile.

What the resize is, in one paragraph each:

* **The lattice is one function.**  `Type.widensTo` has four pairs and
  the two that are missing are the point: `int → float` and
  `long → float` are Java's, and they lose everything above 2^24.
  Across the ladders the answer is always `double`.  `Type.unified`
  is where two operands meet; `fit`, `unifyNumeric`, `widensInto` and
  the list-literal fold all ask one of the two and none decides for
  itself.
* **`ConvertKind` is deleted.**  `convert: Register` goes from the
  operand's type to the register's; the verifier derives which of the
  four families it is, and refuses same-to-same — which is what caught
  a real bug where a folded constant was emitted into a `long`
  register whatever its own width said.
* **A numeric constant is carried at its family's widest member** and
  lands at the register's width, checked exact by the verifier.  The
  language's own rule about literals, one stage down — and the reason
  `byte`, `short` and `half` will add no constant arms.
* **Checked arithmetic at both widths**, one body per family in
  `runtime/operators.zig` and a width parameter through `emitChecked`,
  `checkDivisor` and `correctToFloor`.  An `int` counter squared
  overflows at 46,341 and says so, with a location.
* **A literal lands before it is parsed, never after.**  Three sites
  lowered a value before its place had named a type, and widening
  afterwards is a *different number* for `0.1`: method arguments
  (`methodParameters`, now the one table, consulted by the operand
  batch and by the check), `xs[i] = v`, and binary operands (two
  rules, mirrored in the constant folder so a file-scope `let` and a
  local one agree about what `2 * 0.1` is).
* **`compare_long_double`**, renamed because the old name described a
  function that takes neither an `int` nor a `float`.  Four pairs now
  reach it and it answers all four unchanged; `int` against `double`
  skips the call entirely, as the memo predicted.

The audit fired exactly where it was meant to.  Sixteen sites across
`std/`, `programs/`, `bench/` and the installed-pair test now say
`long` or `double` where they meet a wide value; the ledger and the
reason for each is in *As built*.  A seventeenth kind of loudness was
not a program at all — the builtins compared types exactly where they
had to widen, so `xs[i]` refused an `int` loop counter, and
`widensInto` is now the one place that rule is applied.

Two bugs found on the way, both older than this work: the constant
folder's float `%` used Zig's `@mod` where the runtime floors, so
`let x = 7.0 % -3.0` folded to `1.0` and ran as `-2.0`; and
`emitConstantValue` hard-coded a constant's register width.

`format_version` 20 → 21, and the fingerprint pin with it.

+++ src/luce/07_optimize/test.zig
@@ -295,7 +295,7 @@ fn handBuilt(
     const arena = program.arena.allocator();
 
     const heap_types = try arena.alloc(types.HeapType, 1);
-    heap_types[0] = .{ .list = .int };
+    heap_types[0] = .{ .list = .long };
     program.heap_types = heap_types;
 
     const items = try arena.alloc(Register, instructions.len);
@@ -404,7 +404,7 @@ test "ownership never forwards a load out of a slot that owns its storage" {
     const arena = program.arena.allocator();
 
     const heap_types = try arena.alloc(types.HeapType, 1);
-    heap_types[0] = .{ .list = .int };
+    heap_types[0] = .{ .list = .long };
     program.heap_types = heap_types;
     const layouts = try arena.alloc(types.StructLayout, 1);
     const fields = try arena.alloc(types.StructField, 1);
